Save the job's depend_list in StateSaveLocation.
1. We can know if a dependency was a remote dependency on slurmctld
startup even without the cluster being in a federation. For example, if
the cluster was removed from a federation while it was down, we weren't
able to tell if a dependency was remote or not when the slurmctld
started up again.
2. We don't need to rebuild the dependency list from the string on
startup anymore, because we already have the dependency list.
3. Leave restore_job_dependencies to handle creating depend_list for
previous versions.

Bug 6068