Don't send REQUEST_UPDATE_ORIGIN_DEP to self.
slurmctld shouldn't send an RPC to itself - we do this check in many
other places but weren't doing it here yet.

This can happen if cluster 1 has a job with a dependency on a job in
cluster 2. Cluster 1 sends the dependency to cluster 2. The slurmctld's
are shut down; clusters 1 and 2 are removed from the federation then
re-added to the federation in the opposite order. After starting the
slurmctld's again, cluster 2 continues to test the dependency. When
there's an update for the dependency, it uses the job ID to determine
cluster ID, but the cluster ID's changed, so the job ID may match to
cluster 2 instead of cluster 1. Then cluster 2 would try to send the RPC
to itself, which shouldn't happen, so we prevent that from happening in
this patch.

Bug 6068