Fix seg fault on cons_res when requesting --spread-job.
On cons_res it used to have an array of uint32_t that represented just
avail cpu counts.  When switching it to the array of pointers there
were places missed and there are places where we don't protect against
the pointer not existing before referencing it.  Filling in the array
in this situation makes it behave as before.

Bug 8517


Signed-off-by: Tim Wickberg <tim@schedmd.com>