Docs - Mention that gres.conf is no longer needed for count-only GRES
In 19.05.3 (commit 4d7df8b00, I believe), a change was made in how
slurm.conf and gres.conf were consolidated. A count-only GRES that
could be fully described in slurm.conf no longer needed a gres.conf
entry. Update the docs to point out this convenience.

Other minor doc changes:

* NVML is searched for by default - there is no longer a need to
enable it explicitly. So change from "enabled" to "found."

* In the slurm.conf man page, mention the GRES count limit of 1024
when a type is specified, like it is in the gres.conf man page.

Bug 7905