Fix remote dependency testing for meta job in job array
build_job_queue() sets the array_task_id for the meta job of the job
array when the job array gets split because of a burstbuffer or
correspond dependency. This meant that for local dependencies the meta
job would get tested as the final task of the array with an
array_task_id that was not NO_VAL or INFINITY. However, we weren't
re-sending out the dependency for the meta job to the remote cluster, so
it wasn't getting the array_task_id.

Bug 8574