Remove dead xcpuinfo abs to mac code path when parsing gres.conf Cores.
Logic was initially added in 1f7ce90e3 to translate the parsed abstract
representation of Cores option range to machine/physical so that cpus_bitmap
could later be compared in _get_usable_gres() with the physical CPU
retrieved affinity mask.

A later commit de02f808f9 made it so that slurmd packs and sends the GRES
configuration to the slurmstepd and removes the stepd's call to
gres_plugin_node_config_load(). This function penultimate argument was
the (void *)&xcpuinfo_abs_to_mac pointer used to do the abs to mac
translation in the gres.conf parsing logic, and so after the commit this
became dead code, since this was the only call with non-NULL argument.

This only removes the dead code as a side effect of the mentioned
de02f808f9 change. Since then, cpus_bitmap is abstract according to the
documentation and thus is inconsistent when later compared with the physical
affinity mask interpretation in _get_usable_gres(), and thus this needs to
be addressed somehow.

Bug 9199.