Reset job eligible time when job is manually held.
Making the behavior aligned to that of an independent job that becomes
dependent/ineligible or when a job is directly submitted with -H.

Further cleanup and alignment of actions would be great in other use-cases
of held jobs due to other reasons: manual requeuehold, RequeueExitHold,
HoldMaxRequeue, MAX_BATCH_REQUEUE, invalid burst buffer spec, etc.

Bug 8711.

Signed-off-by: Danny Auble <da@schedmd.com>