Fix EXP_LINE parsing and hashtbl copy functions.
The compiled regex is stored in hashtbl[0], but on copy functions we didn't
copy the regex generating a hashtbl which could lead to a segfault when
calling regexec(). This patch fixes this issue by compiling a new regex
for every new hashtbl - it is not safe to just copy an already compiled
regex since further regfree can affect copied pointers.

This was segfaulting when using EXP_LINE parsing in external code.

Bug 8637.