Use log_flag(STEPS, ...) where able.
Note that a few of these are replacing sched_info() logs. IMO nothing
routed to the sched log should be contingent on the debug flag, thus
changing these over should not cause problems. If they are desired in
the SchedLog, then they should change to a sched_debug() not contingent
on the DebugFlags.