Change log_flag() to log at verbose level and prepend flag name.
Logging at the verbose() level keeps messages out of the user commands
by default (unless -v was requested), which minimizes disruption from
adding additional debugging messages in to the network and RPC layers.

Prepending the debug flag helps identify which messages have been enabled
by a given flag, which is useful when tracing a given subsystem. A log of
associated messages have given hints to this, but making this consistent
with associated cleanup (to come) should simplify tracing when necessary.

Bug 8227.