Send registration message upon resuming down or future down.
A network issue causes a node to be set to down and batch jobs on it can
be requeued. When the node is resumed with:

scontrol update nodename=<node> state=resume

send a registration message (instead of just a ping message) to the node
to clean up jobs or steps that were previously running on the node.

Before this patch, a registration message could already be forced to be
sent by restarting slurmctld or slurmd.

Bug 7831