sacct - Get UID from db to save potentially slow call to getpwnam.
sacct previously used getpwnam() to get the UID given a user name.
However, the UID is stored in the database, so to remove potentially
slow calls to getpwnam() we use the UID from the database instead.

To use the previous behavior of calling getpwnam() to get the UID, we
added the --use-local-uid flag to sacct. This may be desirable if the
UID of the user name in the database doesn't match the UID of the user
name on the node running sacct.

Bug 8263