Testsuite - Fix test5.11 false positive when NodeHostName!=NodeName
The sinfo listing nodes makes use of %n which is NodeHostName the value
is then passed to `scontrol create partition nodes=....` which will
fail since it expects NodeName.

Sinfo comparing CPUs is done in a partition mode, which can lead to
number of CPUs being displayed like NUMBER+ for nodes with the the name
hostname or as name[01-10] with node name expansion. Changing to per
node output (-N).

Bug 8822.