Testsuite - Improve wait_for_job checking for all DONE states
Previously if a job was ended for example with OUT_OF_MEMORY and we call
wait_for_job DONE, the function will keep pulling until
max_job_state_delay.
Now all DONE state are checked, in order.
NOT_FOUND is not a real state, but manually created in the function.

Bug 8837

Signed-off-by: Michael Hinton <hinton@schedmd.com>