Fix MPS without File with 1 GPU and without GPUs
Previously slurmd was discarding MPS if defined without File and the
node had only 1 GPU.
Now we link the MPS to that GPU by default.

Previously slurmd was sending a MPS wihtout GPU RPC to the controller
if that was set in conf files, and the controller was aborting due
internal assertions.
Now even if broken RPC is sent the controller won't abort but return
Invalid Argument, and slurmds are no more sending broken RPCs but fatal
with a clear message instead.

Bug 8775

Co-authored-by: Michael Hinton <hinton@schedmd.com>