Always pack state_reason, not state_reason_prev.
This is a partial revert of commit 74184e4db3b. From that commit message:

"At the start of a scheduling cycle, the job's "reason" field can
be cleared. If the scheduler fails to reach that job and set its
value to a new reason, the original reason was lost and the state
reports would report NoReason."

This isn't true anymore so we revert that change (packing the previous
state_reason) and just pack the job's state_reason.

Bug 8848