slurmdbd - Limit number of jobs updated by as_mysql_flush_jobs_on_cluster()
If there are too many jobs that need to be flushed, then
as_mysql_flush_jobs_on_cluster() can cause too large of a SQL request
that will kill the persistent connection. Upon restart, the same issue
will repeat.

This patch limits the number of jobs to update at once. If the full
limit of jobs are being updated, it will requery to ensure no more jobs
need to be flushed. The requery may be slow but should be rare.

Bug 8176.