Allow the AutoDetect gres.conf option to be node-local
Allow for AutoDetect to be specified on some nodes while not on
others. This allows for a mix of nodes with and without NVML/RSMI
installed while only needing a single gres.conf.

Also change GRES_AUTODETECT_NONE to represent an explicitly-specified
AutoDetect=off, and change GRES_AUTODETECT_UNSET to represent an
unspecified AutoDetect. This helps fall back to the global value when
the node-local value isn't specified.

Bug 8362

Co-authored-by: Danny Auble <da@schedmd.com>