Docs - Add note about Prolog/EpilogSlurmctld to high_throughput.shtml
Both PrologSlurmctld and EpilogSlurmctld shouldn't be used in
high-throughput environment. When enabled they create a separate thread
to hanlde script execution. Handling errors, completition and building
script environment has to be done under a job write lock. In HTC
workload this may lead to thousands of slurmctld threads competing for
the lock.

Bug 7946.