nss_slurm - avoid possible deadlock in _stepd_connect_legacy().
Add stepd_connect_nss to avoid possible deadlocking when writing to the
stepd fails. This version omits the problematic fall-back to
_stepd_connect_legacy() which requires packing an auth credential, which
requires loading the config, which results in a deadlock in slurmstepd.

This has only been seen on overloaded compute nodes, as otherwise the odds
of the first connection failing but the second (in _stepd_connect_legacy())
succeeding are quite low.

This patch is not required for 20.11 as _stepd_connect_legacy() has been
removed.

Bug 9032.