Fix possible double accounting in jobacct_gather
Some return codes were ignored in common_jag.c which could've caused thread
stats to be accounted individually as threads increasing the used resources
of the ppid. This case was rare and of a low impact and usually came with
_is_a_lwp error in stepd logs.

A similar situation happened when NoShare was set and the process was gone
when reading /proc stats, the prec was still taken into account but it
must be discarded since it has possibly incomplete information.

Bug 9093