Fix AMD GPU ROCM 3.5 support.
The rocm_smi.h file started including other header files in this release.

/opt/rocm/rocm_smi/include/rocm_smi/rocm_smi.h:55:10: fatal error:
rocm_smi/kfd_ioctl.h: No such file or directory
   55 | #include "rocm_smi/kfd_ioctl.h"
      |          ^~~~~~~~~~~~~~~~~~~~~~

So it is clear they want the include directory to be
'/opt/rocm/rocm_smi/include' instead of
'/opt/rocm/rocm_smi/include/rocm_smi' as we had it.

All this does is move the include directory up one level and then add
that internal directory wherever we needed to include the header.

This was tested on rocm 3.3 and 3.5.

Bug 9216