Set burst_buffer_state in a job whenever burst buffer state changes.
char *burst_buffer_state;

was already an existing variable in job_record_t. It was already used in
many places. Examples include packing/unpacking it in slurm_protocol_pack,
sending it to squeue, reading it via the rest API with slurmrestd, and
saving it to/loading it from StateSaveLocation. However, it was never
actually assigned. This patch sets this variable to an appropriate
string whenever the job's burst buffer state changes. This has the
effect of saving the state of the burst buffer whenever slurmctld is
shutdown, since burst_buffer_state is saved in StateSaveLocation.
This is needed for the next commit.

Bug 9193