Make x11 forwarding setup async-signal-safe.
In slurmstepd we did a fork to setup X11 forwarding. The fork was done
after stepd became multithreaded and there were many non-reentrant or
non-async-safe calls. This lead to undefined behaviors when some X11
application terminated unexpectedly, possibly killing the extern step by
SIGPIPE or others causing further X11 communications to not work in the
allocation.

This happened reliably with certain applications such as VTune and
Totalview. The approach here removes the fork completely since it was
only required by the original libssh2 X11 forwarding code. So this
removes the dangers of fork() without exec().

We create also a new signal handler thread to properly do the X11
cleanup.

Bug 9289.