Testsuite - Add testsuite_shared_dir, test_dir and _test_init/fini
A new testsuite_shared_dir defined the system dir that the testsuite will
assume that is shared across the cluster.
A temporary test_dir is created for each test in a new _test_init function
that is called automatically when the test begins.
A new _test_fini is now call at the end of a test to remove test_dir
based on the test status and testsuite_cleanup_on_failure.

All print_{header,success,failure,skipped} are now private.
The set test_id and the calls from print_header has been removed from
individual tests.

Bug 8743