Testsuite - Rearrange get_nodes, get_node_param and get_nodes_by_request
Now get_nodes returns a dict of nodes and their configuration parameters,
and get_node_param a specific value of a parameter of one node.
Finally, get_node_by_request returns a list of nodes that can be easily
use as input of the get_nodes to obtain their config, or even as input
of get_node_param if it's one node.

All impacted test are refactored.

Bug 8781