Fix gres_plugin_job_core_filter3 for non-uniform gres placement
Using continue here means that cores from this socket are not considered
available for the job. This is correct only if we are in first_pass ||
enforce_binding or we don't have any cores available on the socket. This
case is handled by if the following if statement.

If we're not enforcing_binding we can pick cores from the socket even if
we don't have any/enough GRES on this socket.

Bug 9471