Cleanup remaining "Processing RPC ..." messages in slurmctld.
Messages that include further details now have "details:" in the message.
Do not bother to print the uid which is now printed at debug2() in
slurmctld_req(). Move these all to debug3() or keep contingent on a flag.

Move the two sched_debug2() calls to sched_debug3() which will not impact
their visibility in the sched log, but ensure they do not print at debug2
alongside the not-common message. Leave them be otherwise in case someone
needs to find them in the sched log.

Drop the two STEPS flagged messages that duplicate the debug2() message.

For the REQUEST_RECONFIGURE message, make this contingent on an authorized
user, and then change the message format to make it distinct. As reconfig
may cause issues (e.g., if the underlying config files are missing changes
that were made through scontrol) we always want a clear high-level message
for it.

And always use %u for uid formatting.