Fix sbcast --fanout option.
The call to slurm_set_tree_width() was meant to handle this, but does not
since the route plugin is internally caching the tree_width value at
route_init() time. So the call to route_g_split_hostlist() will use that
instead of the newly updated value.

Instead, directly set it in the message header to ensure the desired value
is used.

Bug 7741.