Fix slurmctld segfault on reconfigure RPC DDoS
When a very high number of reconfigure RPCs were served by slurmctld,
a segfault could happen if one thread was trying to save the state
while another one was reading the config.

Serving reconfigure RPCs in parallel doesn't make sense, so fix it by
changing it to be sequential instead.

Bug 8680