Change schedule() to use a boolean argument for full_queue.
Arguments were always 0 or INFINITE (indicate a full queue run should be
performed). As such, make the argument easier to understand, and get rid
of some overly-complex intermediate calculations to track this.

Note that as soon as any request has been made for a full queue run, that
request is "sticky" and persists until _schedule() is actually called.

No functional change.