Have NVML report an abstract *Core* affinity range
The abstract range previously reported was a CPU affinity range, and
was not symmetric to the Cores field in gres.conf when threads per
core was > 1. This defeats the whole purpose of giving
xcpuinfo_mac_to_abs() to NVML in the first place, since it only serves
as a gres.conf sanity check (via the logs).

So change xcpuinfo_mac_to_abs() to be the inverse of
xcpuinfo_abs_to_mac(), notably condensing the abstract CPU bitmap
derived from the machine CPU bitmap into an abstract *core* range
string that can be copy and pasted into gres.conf.

Remove _ranges_conv(), as it was used in only one place and the
approach of using bitmaps is more idiomatic than using uint16_t
arrays.

Bug 9211