Change default target file name for "srun --bcast" to include node name.
If "srun --bcast= option ends with '/' it will be used as a directory,
and the destination file name will be slurm_bcast_<job>.<step>.<node>.

To permit the node name to be calculated on each node, the file name
computation needs to be deferred to slurmd/slurmstepd rather than
handled inline in srun as was previously done.

Bug 7736.