Implement RPC queuing within slurmctld.
Each RPC queue operates as a dedicated thread processing a specific RPC
type. Matching RPCs are queued, and when that thread has acquired the
appropriate set of locks all pending RPCs of that type are processed
within that set of locks.

This is designed to reduce lock contention internally by limiting the
number of RPC threads in service at any given time.

RPCs that are not queued are processed as normal.