smail - reduce sleep before seff to 5 seconds
The sleep before seff call is to let the job go to COMPLETED (not
COMPLETING) state and store the information in slurmdbd. 5 seconds
should be enough on healthy installation.

The sleep is async from slurmctld perspective, but it's in the systemd
created cgroup (the slice) of slurmctld and operations like `systemctl
restart slurmctld` will wait till its completition.

Bug 9988