Add a 'tmpfs' job_container plugin
Adds a 'tmpfs' job container plugin that moves jobs
into their own namespace consisting of a private /tmp and a
private /dev/shm. The plugin also provides the ability to have
a namespace.conf that sets the basepath for the plugin to
create the bind mounts as well as an initialization script that
can be used to set up the namespace.

Signed-off-by: Aditi Gaur <agaur@lbl.gov>