Fix cpuset.mems/cpus initialization
In certain systems we've seen cpuset.cpus or cpuset.mems are empty by
default instead of being set when created. This means that when we
init the cpuset cg for a job we leave the value for uid_xxx subdir
empty, which leads to subsequent subdirectories to have these files
also empty, so making the job fail because we cannot attach processes
to a cgroup without mems or cpus assigned, causing a confusing
'no space left' error.

We fix that just by reading the ancestor for uid_xxx and setting the
value too, not assuming it is already set.

Bug 9244