sacctmgr - fix false runaways discarding SHOW_ALL jobs in controller.
Although show runaways is only allowed to >= Operators and they don't
have hidden jobs by default, the logic on _get_runaway_jobs() shouldn't
rely on it but to always load all jobs in the controller.
If at some point we allow show runaways (but not fix them) to random
users, that would avoid false runaways.

Bug 10093