Docs - add cons_tres to NOTE under --distribution
Adding this as completeness. The note cannot fully explain the
complicated logic behind specific cores filtering in cons_tres, that
depends on number of CPUs required to run a job and idle on node.

The goals cons_tres tries to accomplish are:
1) Allocate cores as close as possible to GRES
2) If all cores cannot be allocated on closest sockets, spread the job
on other sockets leaving spaces for other jobs (that may want to use
other GRES on the same sockets).

Bug 10019