_set_gpu_defaults is called only from cons_tres
Remove the if check for is_cons_tres form _set_gpu_defaults and replace
it with xassert, since the function should only be called for cons_tres.

Continuation of commit 0b6faf691c6fb.

Since that was moved into an already gated 'if' for cons_tres no reason
to check it again.

Bug 9947