nss_slurm - link against libslurmfull instead of libslurm
Use libslurmfull since it avoids the slurm_conf_init() call forced as
part of the library init, and nss_slurm specifically does not require
the slurm conf to load.

Without this change, a system without a slurm.conf file will fall back to
hunting for it through the DNS SRV records, and if not found there all
calls into this library will fatal().

Bug 10177.