Fix job getting rejected after being requeued on same node that died
If a node dies while a job is running on it, stale stepd socket files
will still exist. On restart of the node, slurmd would see the file and
add it to the active_job_id account. If the job gets requeued on the
node, it will get rejected as a duplicate jobid. This patch prevents
adding stale stepids to the active_job_id array.

Bug 6773