Move getting usage for a reservation to work off a specific reservation.
It has been realized reservation true usage can't be truncated. Mainly
because the unused_wall is only a number on the reservation record
and not for a specific time period.

The older code could miss some jobs as they were only queried from the
requested time instead of the time the reservation ran for.

It was also discovered if you requested multiple clusters that happen
to have reservations on them of the same id at the same time all
jobs would go on the first reservation instead of correctly per cluster.

What this patch does is fix both these issues by querying jobs per
reservation instead of once for all reservations/clusters.

Bug 10219