Fix false positive oom-kill events on extern step termination.
When using cgroup plugins jobacct_gather_fini tries to remove the step's
memory cgroup directory, generating an event in the cgroup. All the
listeners registered to the eventfd will receive this notification. Since
we cannot differentiate this event from an OOM one, from the extern step
we must read the counter before removing the cgroup hierarchy or we'll get
a false positive.

cgroups v1 notification API says:
'eventfd will be woken up by control file implementation or when the
cgroup is removed.'

Bug 10255