Fix segfault when parsing bad "#SBATCH hetjob" directive
When a hetjob or packjob directive was parsed in an sbatch script, a
segfault could happen if there were non-whitespace characters between
`#SBATCH` and `hetjob` or `packjob`. The fix is to NULL out `argument`,
since it was meant only as a temporary char pointer placeholder here.

Also, `#SBATCH hetjobXXX` could cause an assert. This is because
`argument` was set to point to a preexisting non-xmalloced string, but
was xfreed later. This would have also happened in the cases above if
not for the segfault.

Bug 10305