Kill job if any prolog fails and job is not requeued.
When a prolog fails, slurmd sends a request to slurmctld to requeue the
job. If the requeue doesn't work (the job used --no-requeue or
JobRequeue=0) then slurmd sends a request to slurmctld to complete the
batch job. This can be wrong for a couple of reasons:

(1) The job isn't necessarily a batch job. It could have been started by
salloc or srun.

(2) If a multi-node job had a prolog fail on a node that wasn't the batch
host and requeuing was disabled, when slurmd sends the complete batch
job request to slurmctld, slurmctld ignores that request because it didn't
originate from the batch host. See _slurm_rpc_complete_batch_script().
Therefore, the job isn't killed and runs to completion, even though the
prolog failed and the node was drained.

This commit sends a kill job request instead of a complete batch job
request, ensuring that the job is killed.

Bug 10353