Avoid locking/reading the slurm config when handling stray sockets.
This causes issues in nss_slurm, as nss_slurm is not meant to load the
full configuration as that usually leads to a deadlock. So test slurm_conf
directly and inline the check.

Note that this check was also subtly incorrect - SlurmdUser not SlurmUser
is what matters here for permissions. So fix that as well.

Bug 10405.

Co-authored-by: Tim McMullan <mcmullan@schedmd.com>