Fix cgroup namespace detection when running in a container
When using containers (such as linux containers or docker), slurmds won't
start because when they call xcgroup_ns_is_available to check that they are
in a cgroup, it looks for the presence of the release_agent file. However,
the release_agent file is unique in that it is only present in the root
cgroup. Since containers are namespace-bound to a container-specific child
subdirectory in each control group, it fails to find release_agent and the
slurmd will fail to start.

The fix relies on checking a 'tasks' file instead of 'release_agent'.

Bug 10468