Use new specific spank_has_epilog() function
Continuation of previous commit.

spank_plugin_count() was added in 246ccd109e. With the addition of
spank_has_epilog() in the previous commit, we can use that to determine
if an epilog will run or not instead. This is true to the original
commit's intention c57ab123e9c.

Bug 10456

Signed-off-by: Marcin Stolarek <cinek@schedmd.com>