Change the default srun step allocation behavior to --whole.
This is the way it worked in versions prior to 20.11.

This assigns the entire amount of non-gres resources in the job allocation
on each node to the step.

Add --exact option to get only those resources requested from the user.
(Similar to --exclusive, but does not also impact job allocation behavior.)

Remove documentation for --whole, which is now unnecessary, but left in
place for sites who have already modified their job scripts or started
setting the SLURM_WHOLE environment variable.

Bug 10586.

Signed-off-by: Tim Wickberg <tim@schedmd.com>