Fix job rejection when --gres is less than --gpus
Make sure max_gres is properly reset on each iteration in the while
loop, or else the previous sock_gres's max_gres could erroneously be
used for a subsequent sock_gres.

Also fix a similar potential use-before-reset bug for rem_gres.

The following was being rejected:
    srun --gres=scratch:3 --gpus=2 hostname
While the following ran fine:
    srun --gres=scratch:2 --gpus=2 hostname

The non-GPU GRES would have to be <= gpus for the job to run. This
dependence was a bug.

Bug 10569