jobacct_gather/linux - Don't count memory on tasks that have already finished.
Initialize TRES on all precs in prec_list.

From d86339dc27ed1b9 we never remove a prec from prec_list because
we don't want to loose the cpu time gathered there.  But we don't
want to track that TRES after the process is done.  What this does is
initializes the TRES back to INFINITE so we don't keep track of it
as if the prec is found again on the poll we just replace it leaving
only the precs that have finished in this inited only state.

Bug 10538