Fix salloc/srun sometimes not getting notified on prolog failure.
slurmd relied on ESLURM_BATCH_ONLY being returned to know if it was a
salloc/srun job and sent a message to the job's stdout about prolog
failure. However, in _job_requeue_op(), a few conditions for returning
ESLURM_DISABLED happen before the condition for returning ESLURM_BATCH_ONLY
so this isn't reliable. Instead this commit always sends the message to
the job's stdout about a prolog failure (if there was a prolog failure)
even to batch jobs.

Bug 10353