Print extern step stats in sstat
Unlike regular steps, the extern step has only one pid set to -1.
The mechanism to gather the stats of a step calls jobacct_gather for a
specific pid to get the related statistics. For a regular step each task
has one associated pid.

This patch deals differently with extern step, just calling proctrack
to get all the pids of the extern by using its container id, which is
no more than the slurmstepd.extern pid.

After we have all the pids we can just follow the normal path getting
all the statistics. The only other difference is we need to force the
number of tasks to 1.

Bug 10822