Fix slurmd regression in 20.11 finding/sending the parent node address.
Introduced in 30af622a52, the patch added a new case in reverse_tree_info()
which filled in parent_rank with -1 when width > nodes. This lead to callers
like _send_slurmstepd_init() to attempt to find the slurm_addr_t of this
node's parent slurmd in the step host list with parent_rank = -1 when not
applicable.

This resulted in misleading errors and messages being logged (even with
malformed node names) and incorrect reverse-tree info sent to the stepd.

Bug 10838.