Retry on zero bytes sent in slurm_send_timeout()
Certain drivers will respond with POLLIN and then proceed to return
send()=0 on a otherwise healthy socket. This patch causes a retry of
poll() in this scenario which will continue with normal operations or
catch if the socket is now in an error state.

Logging is tagged with DEBUG_FLAG_NETWORK since the next call of poll()
will catch any errors and in theory no user should ever need to see
this log message.

Bug 10883