Limit data when checking for a default wckey.
By sending in the local cluster list and changing the query to only
give the default wckey first and then the newest one will always be the
one we just added making it not needed to have FOR UPDATE specified.

This also makes it so we only query non-deleted wckeys.

Bug 10561