Fix bf_hetjob_prio truncating SchedulerParameters in main scheduler
strtok() is meant to run until it returns NULL. This is because it actually
modifies the first argument in the meantime. Since we only called it once
while parsing SchedulerParameters in the main scheduler, it basically
truncated everything after bf_hetjob_prio. The solution is to simply use
xstrncasecmp() instead.

The backfill scheduler is unaffected due to 5b833ff38f.

Bug 10895