Requeue or kill job on a prolog failure when PrologFlags is not set.
Whe PrologFlags is not set, the job prolog runs just before the first
step of the job.

Previously, when PrologFlags was not set, the job was not killed on a
prolog failure. Although the step that triggered the prolog failure
would not run, the job's batch script would continue to run and execute
any remaining steps without re-running the prolog.

Bug 10353