Error on invalid options in AccountingStorageEnforce.
Previously, if any part of the string for AccountingStorageEnforce was
invalid, that part was ignored. This could lead to confusing behavior
due to a small typo. For example:

	AccountingStorageEnforce=limit

This was ignored since "limit" is not a valid option, but "limits" is a
valid option.

Many other parameters in slurm.conf are enforced this way, but not all
are. This commit only enforces valid strings for AccountingStorageEnforce.

Bug 10820.