Prevent potential segfault, clear stray CRON_JOB flag, and emit error().
This can happen if a job from 19.05 set the INVALID_DEPEND job flag and
then ran after upgrading to 20.11 due to the reuse of the bit for the
CRON_JOB flag.

This previous commit fixes this on upgrade, but if a site had already
updated there may still be pending jobs left in with the flag set
waiting to cause a segfault.

Bug 11019.

Co-authored-by: Michael Hinton <hinton@schedmd.com>
Co-authored-by: Steve Ford <fordste5@msu.edu>