Restore burst buffer state on slurmctld startup.
Previously even though burst buffers were state saved, the state of the
burst buffer was always reset to PENDING. This commit gets the state of
the burst buffer when slurmctld was shutdown and then uses that
information to do whatever is needed: certain threads are restarted and
the burst buffer state may be changed, or the burst buffer may just be
cleaned up.

Previously a burst buffer that was state saved was ignored if the job
was completed. This commit removed that logic since the burst buffer may
not have completed stage-out and teardown.

Bug 9193