Fix race condition causing burst buffer not to be state saved.
The burst buffer record which is state saved was not created for a job
until the middle of the stage-in thread. If slurmctld was shutdown during
burst buffer stage-in for a job, the thread would be killed before
creating the burst buffer record, and therefore the burst buffer was
not state saved.

Bug 9193