Minor cleanup - remove unneeded code.
Due to other changes for bug 9193, finding job_ptr here is no longer
needed, so remove it. Because job_ptr isn't used here, we can also move
where we lock the job_write_lock since we don't need it in the else
block anymore, but we still need it in the if block.

Bug 9193