common/fd - fix getsockopt() call in fd_get_socket_error().
The `errlen` variable was incorrectly assigned the size of the pointer
instead of the buffer pointed to by the pointer before being handed as
last `optlen` value-result argument to getsockopt().

On the other hand, the penultimate `optval` argument was passed as a
pointer pointer instead of just a pointer.

Bug 11098.

Co-authored-by: Alejandro Sanchez <alex@schedmd.com>