_valid_features return NULL if no node matched
If _valid_features didn't find any matching nodes returning NULL will
jump/continue in a loop over nodesets - there is no need to perform
other checks in the main loop of _build_node_list, since we already know
that this nodeset doesn't have any of XOR required features.

It looks like this was the way it was before ef27e18392 and this
was just lost when things were converted to bitmaps in that commit.

Bug 9578

Co-authored-by: Marcin Stolarek <cinek@schedmd.com>