Docs - Update info on --gpu-bind
Point out that by default, tasks can access all GPUs in the *step*,
rather than the *job*, since this is different when devices are
constrained via task/cgroup and the step only uses a subset of the
GPUs allocated to the job.

Also mention that --gpu-bind is completely ignored if -n1. GPU task
binding does not make any sense if there is only one task in the step;
in that case, you would have just requested less GPUs for the single
task to begin with.

Bug 11226