slurmrestd - openapi.json, remove users parameter from 'slurmdb/jobs/'
The libslurm API doesn't support users query by list of user names as was
advertised by the schema.

As we don't want slurmrestd to handle UIDs, and cannot assume the
slurmrestd process can correctly handle resolving usernames to UIDs, we are
simply removing this (non-functional) field from the schema for now.

Bug 11330.