Make _rm_data able to ignore a file removal failure.
Then, make _rm_data call in _create_ns keep old behavior, as
here it will only be removed if something fails in creation, and no
previous NS leak is possible. So force removal, or fail.

But for _delete_ns, it could be called at job's end after slurmd
got killed and restarted, thus having leaked the NS. Even though,
slurmd can recreate the NS and mount it in the same place, the .ns
file can't be removed at the end of the job because EBUSY.

Bug 11093