Fix single Port= not being applied to range of nodes in slurm.conf
Fix regression from a2678ab5560.

The original intent of a2678ab5560 was to set port in node structure
from the SlurmdPort when nothing was given in NodeName line.
Unfortunately it changed the behavior for multiple nodes defined in one
line with only a single Port value, like: NodeName=test[01-02] Port=3022
will set test01 on 3022, but test02 on SlurmdPort.

Bug 11384